refactor: split big code chunks and rearrange files
