+++ src/assets/styles/style.scss
@@ -122,7 +122,8 @@ $hover-key-bg: #FE83C6;
     display: inline-block;
   }
 
-  &.shift {
+  &.shift-left,
+  &.shift-right {
       .eng-up{
         display: inline-block;
     }
@@ -143,7 +144,8 @@ $hover-key-bg: #FE83C6;
     display: block;
   }
 
-  &.shift {
+  &.shift-left,
+  &.shift-right {
     .ru-up {
       display: block;
     }
